requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
attrs==24.2.0
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
datasets==3.0.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.0
fastapi==0.115.3
filelock==3.16.1
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
instructor==1.6.3
Jinja2==3.1.4
jiter==0.5.0
jsonpath-python==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==2.1.2
oauthlib==3.2.2
openai==1.52.2
packaging==23.2
pandas==2.2.3
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.9.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
tenacity==9.0.0
text-unidecode==1.3
tqdm==4.66.5
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.1
xxhash==3.5.0
yarl==1.16.0
